+++ test/chain.test.js
@@ -62,7 +62,7 @@ describe( 'Chain component', () => {
 
 } );
 
-describe( 'Chain component', () => {
+describe( 'InnerChain component', () => {
     afterEach( cleanup );
     const Context = () => (
         <ChainContext.Consumer>
@@ -70,46 +70,133 @@ describe( 'Chain component', () => {
                 <React.Fragment>
                     <p className={ `${ context.id }-id` }>{ context.id }</p>
                     <p className={ `${ context.id }-state` }>{ context.state }</p>
+                    <p className={ `${ context.id }-parent-state` }>{ context.parentState || 'null' }</p>
+                    { map( context.innerChains, ( state, id ) => (
+                        <div id={ `inner-${ id }` } key={ id }>
+                            <p className="inner-chain-state">{ state }</p>
+                        </div>
+                    ) ) }
                 </React.Fragment>
             ) }
         </ChainContext.Consumer>
     );
 
-    it( 'transitions to \'exited\' on the Outer chain but transitions to \'entered\' state on InnerChain', async () => {
+    it( 'Enters parent-first and exits parent-first', async () => {
         const { container, rerender } = render(
-            <Chain id="test1" timeout={ 45 } appear in={ false }>
+            <Chain id="test3" in timeout={ 500 }>
                 <Context />
-                <InnerChain reverse id="test2">
+                <InnerChain id="test4" timeout={ 1000 }>
                     <Context />
                 </InnerChain>
             </Chain>
         );
 
-        expect( container.querySelector( '.test1-id' ).innerHTML ).toEqual( 'test1' );
-        await wait( () => expect( container.querySelector( '.test1-state' ).innerHTML ).toEqual( 'exited' ) );
-        expect( container.querySelector( '.test2-id' ).innerHTML ).toEqual( 'test2' );
-        await wait( () => expect( container.querySelector( '.test2-state' ).innerHTML ).toEqual( 'entered' ) );
-
-        rerender(
-            <Chain id="test1" timeout={ 45 } appear in>
-                <Context />
-                <InnerChain inOnEntering reverse id="test2">
-                    <Context />
-                </InnerChain>
-            </Chain>
-        );
-        await wait( () => expect( container.querySelector( '.test1-state' ).innerHTML ).toEqual( 'entering' ) );
-        expect( container.querySelector( '.test2-state' ).innerHTML ).toMatch( /exiting|exited/ );
+        const parentState = container.querySelector( '.test3-state' );
+        const childState = container.querySelector( '.test4-state' );
+        expect( parentState && childState ).toBeTruthy();
+        await wait( () => expect( parentState.innerHTML ).toMatch( /entered/ ) );
+        expect( childState.innerHTML ).toMatch( /exited|entering/ );
 
         rerender(
-            <Chain id="test1" timeout={ 45 } appear in={ false }>
+            <Chain id="test3" timeout={ 500 }>
                 <Context />
-                <InnerChain inOnEntering reverse id="test2">
+                <InnerChain id="test4" timeout={ 1000 }>
                     <Context />
                 </InnerChain>
             </Chain>
         );
-        await wait( () => expect( container.querySelector( '.test1-state' ).innerHTML ).toEqual( 'exiting' ) );
-        expect( container.querySelector( '.test2-state' ).innerHTML ).toMatch( /entering|entered/ );
+
+        await( () => expect( parentState.innerHTML ).toMatch( /exited/ ) );
+        expect( childState.innerHTML ).toMatch( /entering/ );
+    } );
+
+    it( 'Enters parent-first and exits parent-last', async () => {
+        const { container, rerender } = render(
+            <Chain id="test3" order="filo" in timeout={ 500 }>
+                <Context />
+                <InnerChain id="test4" timeout={ 1000 }>
+                    <Context />
+                </InnerChain>
+            </Chain>
+        );
+
+        const parentState = container.querySelector( '.test3-state' );
+        const childState = container.querySelector( '.test4-state' );
+        expect( parentState && childState ).toBeTruthy();
+        await wait( () => expect( parentState.innerHTML ).toMatch( /entering|entered/ ) );
+        expect( childState.innerHTML ).toMatch( /exited/ );
+        await wait( () => expect( childState.innerHTML ).toMatch( /entered/ ) );
+
+        rerender(
+            <Chain id="test3" order="filo" timeout={ 500 }>
+                <Context />
+                <InnerChain id="test4" timeout={ 1000 }>
+                    <Context />
+                </InnerChain>
+            </Chain>
+        );
+
+        expect( childState.innerHTML ).toMatch( /exiting/ );
+        expect( parentState.innerHTML ).toMatch( /entered/ );
+    } );
+
+    it( 'Enters parent-last and exits parent-first', async () => {
+        const { container, rerender } = render(
+            <Chain id="test3" order="lifo" in timeout={ 1000 }>
+                <Context />
+                <InnerChain id="test4" timeout={ 500 }>
+                    <Context />
+                </InnerChain>
+            </Chain>
+        );
+
+        const parentState = container.querySelector( '.test3-state' );
+        const childState = container.querySelector( '.test4-state' );
+        expect( parentState && childState ).toBeTruthy();
+        expect( childState.innerHTML ).toMatch( /entering/ );
+        expect( parentState.innerHTML ).toMatch( /exiting|exited/ );
+        await wait( () => expect( parentState.innerHTML ).toMatch( /entered/ ) );
+
+        rerender(
+            <Chain id="test3" order="lifo" timeout={ 1000 }>
+                <Context />
+                <InnerChain id="test4" timeout={ 500 }>
+                    <Context />
+                </InnerChain>
+            </Chain>
+        );
+
+        expect( parentState.innerHTML ).toMatch( /exiting/ );
+        expect( childState.innerHTML ).toMatch( /entered|exiting/ );
+    } );
+
+    it( 'Enters parent-last and exits parent-last', async () => {
+        const { container, rerender } = render(
+            <Chain id="test3" order="lilo" in timeout={ 1000 }>
+                <Context />
+                <InnerChain id="test4" timeout={ 500 }>
+                    <Context />
+                </InnerChain>
+            </Chain>
+        );
+
+        const parentState = container.querySelector( '.test3-state' );
+        const childState = container.querySelector( '.test4-state' );
+        expect( parentState && childState ).toBeTruthy();
+        expect( childState.innerHTML ).toMatch( /entering/ );
+        expect( parentState.innerHTML ).toMatch( /exiting/ );
+        await wait( () => expect( parentState.innerHTML ).toMatch( /entered/ ) );
+
+        rerender(
+            <Chain id="test3" order="lilo" timeout={ 1000 }>
+                <Context />
+                <InnerChain id="test4" timeout={ 500 }>
+                    <Context />
+                </InnerChain>
+            </Chain>
+        );
+
+        expect( childState.innerHTML ).toMatch( /exiting/ );
+        expect( parentState.innerHTML ).toMatch( /entered/ );
     } );
 } );